Import helper functions from load_question
